client/requirements.txt
websocket-client==0.57.0
redis==3.5.3
Flask==1.1.2
Flask-Compress==1.4.0
Flask-Executor==0.9.3
Flask-RESTful==0.3.8
Flask-SocketIO==2.9.6
